+++ CHANGELOG.rst
@@ -27,7 +27,7 @@ Added
     - ``exclude_inherited`` argument in ``Project.get_delegates()`` (#595)
     - Optional options for app settings of type ``STRING`` and ``INTEGER`` (#592)
     - Display placeholders for app setting form fields (#584)
-    - Support for local user invites (#548, #613, #615)
+    - Support for local user invites (#548, #613, #615, #621)
     - Local user account creation and updating (#547)
     - ``batchupdateroles`` management command (#15, #602)
     - Project invite REST API views (#15, #598)